Change project name to kmp-process
