+++ gradle.properties
@@ -12,14 +12,14 @@ kotlin.native.ignoreDisabledTargets=true
 SONATYPE_HOST=S01
 RELEASE_SIGNING_ENABLED=true
 
-GROUP=io.matthewnelson.process
+GROUP=io.matthewnelson.kmp-process
 
 POM_INCEPTION_YEAR=2024
 
-POM_URL=https://github.com/05nelsonm/process/
-POM_SCM_URL=https://github.com/05nelsonm/process/
-POM_SCM_CONNECTION=scm:git:git://github.com/05nelsonm/process.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/05nelsonm/process.git
+POM_URL=https://github.com/05nelsonm/kmp-process/
+POM_SCM_URL=https://github.com/05nelsonm/kmp-process/
+POM_SCM_CONNECTION=scm:git:git://github.com/05nelsonm/kmp-process.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/05nelsonm/kmp-process.git
 
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=https://www.apache.org/licenses/LICENSE-2.0.txt
