+++ library/process/gradle.properties
@@ -1,3 +1,3 @@
 POM_ARTIFACT_ID=process
-POM_NAME=Process
-POM_DESCRIPTION=Process
+POM_NAME=Kotlin Multiplatform Process
+POM_DESCRIPTION=Process for Kotlin Multiplatform
