+++ library/process/src/nativeInterop/cinterop/spawn.def
@@ -1,3 +1,3 @@
-package = io.matthewnelson.process
+package = io.matthewnelson.kmp.process.internal
 headers = spawn.h
 headerFilter = spawn.h
