+++ library/process/src/nativeMain/kotlin/io/matthewnelson/kmp/process/internal/NativePlatform.kt
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2024 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("KotlinRedundantDiagnosticSuppress")
+
+package io.matthewnelson.kmp.process.internal
+
+import io.matthewnelson.kmp.file.DelicateFileApi
+import io.matthewnelson.kmp.file.IOException
+import io.matthewnelson.kmp.file.errnoToIOException
+import kotlinx.cinterop.ExperimentalForeignApi
+import platform.posix.errno
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.InvocationKind
+import kotlin.contracts.contract
+
+@Suppress("NOTHING_TO_INLINE")
+internal expect inline fun PlatformBuilder.parentEnvironment(): MutableMap<String, String>
+
+@Suppress("NOTHING_TO_INLINE")
+@Throws(IOException::class)
+@OptIn(ExperimentalContracts::class)
+internal inline fun Int.check(
+    block: (result: Int) -> Boolean = { it >= 0 },
+): Int {
+    contract {
+        callsInPlace(block, InvocationKind.EXACTLY_ONCE)
+    }
+
+    if (!block(this)) {
+        @OptIn(DelicateFileApi::class, ExperimentalForeignApi::class)
+        throw errnoToIOException(errno)
+    }
+
+    return this
+}
