+++ library/process/src/unixMain/kotlin/io/matthewnelson/kmp/process/internal/PosixSpawnAttrs.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2024 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+
+package io.matthewnelson.kmp.process.internal
+
+import io.matthewnelson.kmp.file.IOException
+import kotlinx.cinterop.CValuesRef
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.MemScope
+
+@OptIn(ExperimentalForeignApi::class)
+internal expect value class PosixSpawnAttrs private constructor(
+    private val _ref: CValuesRef<*>,
+) {
+
+    internal companion object {
+
+        @Throws(IOException::class)
+        internal fun MemScope.posixSpawnAttrInit(): PosixSpawnAttrs
+    }
+}
